backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
pydantic==2.6.4
easypost==9.4.0
beautifulsoup4==4.12.3
